added: type-mixin to community and post models
